add paymentController

+++ rentACar/src/main/java/com/rentACar/rentACar/services/abstracts/CarService.java
@@ -15,5 +15,5 @@ public interface CarService {
     void delete(int id);
     boolean controlCarId(int id);
     int carKilometer(int id);
-    double carDailyPrice(int id);
+    Float carDailyPrice(int id);
 }
